+++ docs/source/pydrobert.torch.rst
@@ -44,14 +44,6 @@ pydrobert.torch.training module
    :undoc-members:
    :show-inheritance:
 
-pydrobert.torch.util module
----------------------------
-
-.. automodule:: pydrobert.torch.util
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 
 Module contents
 ---------------
