+++ src/pydrobert/torch/_lm.py
@@ -15,137 +15,11 @@
 import abc
 import re
 
-from typing import Any, Dict, Optional, Sequence, TextIO, Tuple, TYPE_CHECKING, Union
+from typing import Any, Dict, Optional, Sequence, Tuple, TYPE_CHECKING, Union
 
 import torch
 
 from ._compat import script
-
-
-def parse_arpa_lm(file_: Union[TextIO, str], token2id: Optional[dict] = None) -> list:
-    r"""Parse an ARPA statistical language model
-
-    An `ARPA language model <https://cmusphinx.github.io/wiki/arpaformat/>`__
-    is an n-gram model with back-off probabilities. It is formatted as::
-
-        \data\
-        ngram 1=<count>
-        ngram 2=<count>
-        ...
-        ngram <N>=<count>
-
-        \1-grams:
-        <logp> <token[t]> <logb>
-        <logp> <token[t]> <logb>
-        ...
-
-        \2-grams:
-        <logp> <token[t-1]> <token[t]> <logb>
-        ...
-
-        \<N>-grams:
-        <logp> <token[t-<N>+1]> ... <token[t]>
-        ...
-
-        \end\
-
-    Parameters
-    ----------
-    file_ : str or file
-        Either the path or a file pointer to the file
-    token2id : dict, optional
-        A dictionary whose keys are token strings and values are ids. If set,
-        tokens will be replaced with ids on read
-
-    Returns
-    -------
-    prob_list : list
-        A list of the same length as there are orders of n-grams in the
-        file (e.g. if the file contains up to tri-gram probabilities then
-        `prob_list` will be of length 3). Each element is a dictionary whose
-        key is the word sequence (earliest word first). For 1-grams, this is
-        just the word. For n > 1, this is a tuple of words. Values are either
-        a tuple of ``logp, logb`` of the log-probability and backoff
-        log-probability, or, in the case of the highest-order n-grams that
-        don't need a backoff, just the log probability.
-    
-    Warnings
-    --------
-    This function is not safe for JIT scripting or tracing.
-    """
-    if isinstance(file_, str):
-        with open(file_) as f:
-            return parse_arpa_lm(f, token2id=token2id)
-    line = ""
-    for line in file_:
-        if line.strip() == "\\data\\":
-            break
-    if line.strip() != "\\data\\":
-        raise IOError("Could not find \\data\\ line. Is this an ARPA file?")
-    ngram_counts = []
-    count_pattern = re.compile(r"^ngram\s+(\d+)\s*=\s*(\d+)$")
-    for line in file_:
-        line = line.strip()
-        if not line:
-            continue
-        match = count_pattern.match(line)
-        if match is None:
-            break
-        n, count = (int(x) for x in match.groups())
-        if len(ngram_counts) < n:
-            ngram_counts.extend(0 for _ in range(n - len(ngram_counts)))
-        ngram_counts[n - 1] = count
-    prob_list = [dict() for _ in ngram_counts]
-    ngram_header_pattern = re.compile(r"^\\(\d+)-grams:$")
-    ngram_entry_pattern = re.compile(r"^(-?\d+(?:\.\d+)?)\s+(.*)$")
-    while line != "\\end\\":
-        match = ngram_header_pattern.match(line)
-        if match is None:
-            raise IOError('line "{}" is not valid'.format(line))
-        ngram = int(match.group(1))
-        if ngram > len(ngram_counts):
-            raise IOError(
-                "{}-grams count was not listed, but found entry" "".format(ngram)
-            )
-        dict_ = prob_list[ngram - 1]
-        for line in file_:
-            line = line.strip()
-            if not line:
-                continue
-            match = ngram_entry_pattern.match(line)
-            if match is None:
-                break
-            logp, rest = match.groups()
-            tokens = tuple(rest.strip().split())
-            # IRSTLM and SRILM allow for implicit backoffs on non-final
-            # n-grams, but final n-grams must not have backoffs
-            logb = 0.0
-            if len(tokens) == ngram + 1 and ngram < len(prob_list):
-                try:
-                    logb = float(tokens[-1])
-                    tokens = tokens[:-1]
-                except ValueError:
-                    pass
-            if len(tokens) != ngram:
-                raise IOError(
-                    'expected line "{}" to be a(n) {}-gram' "".format(line, ngram)
-                )
-            if token2id is not None:
-                tokens = tuple(token2id[tok] for tok in tokens)
-            if ngram == 1:
-                tokens = tokens[0]
-            if ngram != len(ngram_counts):
-                dict_[tokens] = (float(logp), logb)
-            else:
-                dict_[tokens] = float(logp)
-    if line != "\\end\\":
-        raise IOError("Could not find \\end\\ line")
-    for ngram_m1, (ngram_count, dict_) in enumerate(zip(ngram_counts, prob_list)):
-        if len(dict_) != ngram_count:
-            raise IOError(
-                "Expected {} {}-grams, got {}".format(ngram_count, ngram_m1, len(dict_))
-            )
-    return prob_list
 
 
 class SequentialLanguageModel(torch.nn.Module, metaclass=abc.ABCMeta):
