+++ tests/test_lm.py
@@ -14,7 +14,6 @@
 
 import os
 import itertools
-from tempfile import SpooledTemporaryFile
 
 from typing import Dict, Tuple
 
@@ -22,7 +21,7 @@ import torch
 import pytest
 
 from pydrobert.torch.modules import LookupLanguageModel, MixableSequentialLanguageModel
-from pydrobert.torch.util import parse_arpa_lm
+from pydrobert.torch.data import parse_arpa_lm
 
 INF = float("inf")
 NAN = float("nan")
@@ -108,115 +107,6 @@ class RNNLM(MixableSequentialLanguageModel):
             torch.nn.functional.log_softmax(logits, -1),
             {"hidden": h_1, "cell": c_1},
         )
-
-
-@pytest.mark.cpu
-def test_parse_arpa_lm():
-    file_ = SpooledTemporaryFile(mode="w+")
-    file_.write(
-        r"""\
-This is from https://cmusphinx.github.io/wiki/arpaformat/
-I've removed the backoff for </s> b/c IRSTLM likes to do things like that
-
-\data\
-ngram 1=7
-ngram 2=7
-
-\1-grams:
--1.0000 <unk>	-0.2553
--98.9366 <s>	 -0.3064
--1.0000 </s>
--0.6990 wood	 -0.2553
--0.6990 cindy	-0.2553
--0.6990 pittsburgh		-0.2553
--0.6990 jean	 -0.1973
-
-\2-grams:
--0.2553 <unk> wood
--0.2553 <s> <unk>
--0.2553 wood pittsburgh
--0.2553 cindy jean
--0.2553 pittsburgh cindy
--0.5563 jean </s>
--0.5563 jean wood
-
-\end\
-"""
-    )
-    file_.seek(0)
-    ngram_list = parse_arpa_lm(file_)
-    assert len(ngram_list) == 2
-    assert set(ngram_list[0]) == {
-        "<unk>",
-        "<s>",
-        "</s>",
-        "wood",
-        "cindy",
-        "pittsburgh",
-        "jean",
-    }
-    assert set(ngram_list[1]) == {
-        ("<unk>", "wood"),
-        ("<s>", "<unk>"),
-        ("wood", "pittsburgh"),
-        ("cindy", "jean"),
-        ("pittsburgh", "cindy"),
-        ("jean", "</s>"),
-        ("jean", "wood"),
-    }
-    assert abs(ngram_list[0]["cindy"][0] + 0.6990) < 1e-4
-    assert abs(ngram_list[0]["jean"][1] + 0.1973) < 1e-4
-    assert abs(ngram_list[1][("cindy", "jean")] + 0.2553) < 1e-4
-    file_.seek(0)
-    token2id = dict((c, hash(c)) for c in ngram_list[0])
-    ngram_list = parse_arpa_lm(file_, token2id=token2id)
-    assert set(ngram_list[0]) == set(token2id.values())
-    file_.seek(0)
-    file_.write(
-        r"""\
-Here's one where we skip right to 10-grams
-
-\data\
-ngram 10 = 1
-
-\10-grams:
-0.0 1 2 3 4 5 6 7 8 9 10
-
-\end\
-"""
-    )
-    file_.seek(0)
-    ngram_list = parse_arpa_lm(file_)
-    assert all(x == dict() for x in ngram_list[:-1])
-    assert not ngram_list[9][tuple(str(x) for x in range(1, 11))]
-    file_.seek(0)
-    file_.write(
-        r"""\
-Here's one where we erroneously include backoffs
-
-\data\
-ngram 1 = 1
-
-\1-grams:
-0.0 a 0.0
-
-\end\
-"""
-    )
-    file_.seek(0)
-    with pytest.raises(IOError):
-        parse_arpa_lm(file_)
-    file_.seek(0)
-    file_.write(
-        r"""\
-Here's an empty one
-
-\data\
-\end\
-"""
-    )
-    file_.seek(0)
-    assert parse_arpa_lm(file_) == []
 
 
 @pytest.mark.cpu
